use cast where appropriate; it fixes some portability problems

+++ src/scanner.c
@@ -365,7 +365,7 @@ vh_scanner_run (scanner_t *scanner, int loop, uint16_t timeout, int priority)
   /* if timeout is 0, there is no sleep between loops */
   if (timeout)
   {
-    scanner->timeout = timeout * 1000000000ULL;
+    scanner->timeout = (uint64_t) timeout * 1000000000;
     vh_timer_thread_start (scanner->timer);
   }
